Refactored draw view classes to enable sublassing
To be able to easily create subclasses of Skia views we need to refactor the RNSkDrawView which now contains too much logic for one class.

This change splits the draw view into a base draw view called RNSkView (defined in RNSkView.h) - which is responsible for setting up the draw loop and dispatching redraws whenever there are draw requests waiting.

Rendering to the canvas is now delegated to a base class called RNSkRenderer (defined in RNSkView.h as well). I have included a renderer for rendering using the Javascript onDraw property (RNSkJSView.h).

The canvas is now provided by a class called RNSkCanvasProvider (defined in RNSkView.h as well) that is overridden with a Metal and OpenGL implementation on iOS/Android.

Rendering using the onDraw / JS is defined in the file RNSkJsView.h and should be easy to read and understand. It uses a subclass of the RNSkRenderer to set up and execute the Javascript code on the right threads.

Both iOS/Android have been updated to use these new base classes.

Files have been renamed to fit better with contents of the classes they contain.