Optimized loops in native
This optimization takes the CPU meter from 32-35% usage down to 31-33% on iOS.

When running tight loops we previously copied the contents of the vector we wanted to loop through in two different areas; the RNSkReadonlyValue and the RNSkPlatformContext notifiers. This was done to avoid crash if we were inserting elements at the same time as we were enumerating - and we also had some locking going on in there.

This has now been removed and replaced by a tight loop that does not allocate new elements. The new loops are protected by a lock.