Updated getDerivedValue() to return const and fixed issue with mutation in JsiImageShader node
The Image Shader node was previously set up so that the local matrix was updated multiple times, aggregating errors when the transform prop was not changed.

This mutation is not something we wan't so this PR changes the getDerivedValue() on props to return a readonly value - which can NOT be mutatet. Any mutations must be done by creating a copy locally and modifying it.

This caused a few other props to no longer compile, and these were also updated after this change to avoid any extra mutations.

I tried writing a test for this, but since it required us to control multiple render passes I was not able to get the test runner to deterministically do this correctly.

Instead the Filters example has been updated with touch and state - clicking on the example should no longer cause aggregated errors in the shader.

fixes #1200