Added centralized property updated system to avoid having to enumerate all props in all nodes when doing a render cycle.
What happens now is that whenever a property is updated, it notifies through the _propertyDidUpdate callback - and the root node (the one without a parent will save the node in a double buffered set of changed props.

When the lifecycle updates and resovles prop changes it can now be done directly from the root node without any enumeration, only updating / resolving the properties that was actually changed.