Removed use of std::dynamic_pointer_cast in loops
For large DOM trees the use of std::dynamic_pointer_cast to detect if we have a render or declaration node is not working well - it shows up in the profiler as hot spots.

This commit introduces a simple shortcut so that nodes now implements a way to return the class of node when needed.

Usage of this has been implemented in the most critical loops in the dom renderer.