Added locking to picture access in custom drawing node.
To avoid race conditions between JS and main thread in the custom drawing node we need to protected the shared resource - the SkPicture - to avoid race conditions.

The lock was added to the Node and not the property since it is the node that sets up and handles threading in this case.

fixes Crash on Android fatal error: "assert(this->getRefCnt() > 0)" #1271