Fixed gravity on metal layer to avoid animations on size change
By some strange reason (we've seen this before - and also tried to fix it before) the CAMetalLayer resizes itself in an animated way when changing size.

This is something we don't want, and we've had a few issues regarding this.

This fixes the problem by setting the gravity on the layer to bottom left - this removes the error we're seeing in the reports.

Fixes #716