Update Surface.MakeOffscren() to work on Node
On Node, we use a polyfill of OffscreenCanvas. However when using MakeWebGLCanvasSurface, we're not able to "trick" CanvasKit in thinking that the API exists, this check from CanvasKit will fail: https://github.com/google/skia/blob/main/modules/canvaskit/webgl.js#L169

This is a way to workaround the issue and enables Offscreen GPU on Node.