Suggested changes RuntimeAwareStore
After doing some instrumentation I tried optimising the JsiHostObject a bit to avoid too much use of the new RuntimeAware store. This caused some mismatches with which runtime was registered as the main runtime (secondary runtime became primary etc.)

This commit is a fix for this issue, and added support for multiple secondary runtimes with the same pattern as the current implementation.

- Added registration of primary runtime in the cache
- Added support for multiple runtimes
- Refactored classes / files to be more inline with terminology in RNSkia