Added runOnMainThread and makeScreenshot in platform context
The platform context now has two additional methods, the runOnMainThread and makeScreenshot methods. These will need to be overridden by implementations since they are considered platform services.

runOnMainThread should let you run a function on the main (UI) thread of the application, while the makeScreenshot should take a screenshot of the view represented by a React view tag and return an SkImage of the view and its children.