This fixes issue with header search paths for monorepo
Instead of hardcoding a relative path we now use the $PODS_TARGET_SRCROOT constant which should fix the issue with node_modules being located a level above in mono repos.

I've also tested this with `use_frameworks! linkage: :static` now and it seems to be working as well.

fixes #1477
fixes #652 (again?)