Relaxed renderer bail-out when iOS is inactive
On iOS we need to be careful about using the Metal primitives when the app is backgrounded.

We have a test that bails out from rendering when the app is in the background / inactive, but we believe that this is a bit to strict (see #1625 and originally #1257).

This commit fixes this by removing bailing out when the app is inactive - Metal can still render, and the original error message was showing errors when the app was backgrounded.