Refactored the screenshot mechanism on Android
Refactored to ensure that child views are rendered in order so that we can combine TextureViews with regular views and keep things like opacity, positions etc.

The issue contains a reproduction that contains a few hard cases like transparency between Skia and React views to ensure we can do this correctly.

The changes I've made is to instead of rendering the Android view (including children) first and then the Surface and Texture Views (which basically renders Skia views on top of React views independant of the order) we now render children in order and recursively.

Fixes #1633