Added implementation of onDraw with drawing a bitmap of the Skia texture on first frame
The logic is to do this in the Java View onDraw method:
If we haven't got a surface yet, let's ask the view to
draw into a bitmap and then render the bitmap. This method is typically only called once - for the first frame, and then the surface will be available and all rendering will be done directly to the surface itself.