Updated Android renderer after CodeReview
Simplified the renderer so that we have a SkiaSurfaceProvider class that has two subclasses, one for creating offscreen and one for creating windowed surfaces.

A fix is made to clean the first frame bitmap that I'm not super happy with. The problem here is that the bitmap "sticks" to the view and needs to be cleared so that we can render to the surface again with a changed bitmap containing transparency.

See the problem in the Stickers example before this fix.