Experiment: Android/Change to render on separate thread
This commit changes so that rendering happens on a separate render thread as we've seen in most examples using OpenGL and TextureViews on Android. When measuring frame rate and performance we get better consistently results by doing this.

Rendering is still locked to the current frame rate by using the Coreographers frame callback to dispatch single render messages to the render thread on Android.

This will of course once again introduce the problem with rendering the first frame, but it seems like it will provide a more smooth and consistent frame rate after that. I have not been able to make it crash, and this will also be backwards compatible with earlier devices.