fix(🖼️): fix thread safety in SkImage and viewRef.makeImageSnapshot (#2761)
Add thread boundaries on images that could be used in a different thread than the one they were created on. The solution is to use the graphic context available in the thread instead of using the "saved" within the image.
Skia now has a `SK_IMAGE_READ_PIXELS_DISABLE_LEGACY_API` that helps with catching these (however APIs like `makeImageSnapshot(GrDirectContext* gr = nullptr)` are not caught by the flag).